feat/move-to-main: renderer uses reducers code from server folder
