+++ server/@types/classnames/index.d.ts
@@ -1,0 +1,4 @@
+declare module 'classnames' {
+	function ClassNames (...args: (string | { [key: string]: any })[]): string
+	export = ClassNames
+}
