--- src/reducers/channelActions.ts
@@ -1,8 +1,0 @@
-export const SET_OUTPUT_LEVEL = 'SET_OUTPUT_LEVEL'
-export const SET_AUX_LEVEL = 'SET_AUX_LEVEL'
-export const SET_COMPLETE_CH_STATE = 'SET_COMPLETE_CH_STATE'
-export const FADE_ACTIVE = 'FADE_ACTIVE'
-export const SET_ASSIGNED_FADER = 'SET_ASSIGNED_FADER'
-export const SET_PRIVATE = 'SET_PRIVATE'
-export const SET_OPTION = 'SET_OPTION'
-
