--- src/reducers/faderActions.ts
@@ -1,29 +1,0 @@
-export const SET_VU_LEVEL = 'SET_VU_LEVEL'
-export const SET_COMPLETE_FADER_STATE = 'SET_COMPLETE_FADER_STATE'
-export const SET_FADER_LEVEL = 'SET_FADER_LEVEL'
-export const SET_ALL_VU_LEVELS = 'SET_ALL_VU_LEVELS'
-export const SET_CHANNEL_LABEL = 'SET_CHANNEL_LABEL'
-export const SET_FADER_THRESHOLD = 'SET_FADER_THRESHOLD'
-export const SET_FADER_RATIO = 'SET_FADER_RATIO'
-export const SET_FADER_LOW = 'SET_FADER_LOW'
-export const SET_FADER_MID = 'SET_FADER_MID'
-export const SET_FADER_HIGH = 'SET_FADER_HIGH'
-export const SET_FADER_MONITOR = 'SET_FADER_MONITOR'
-export const TOGGLE_PGM = 'TOGGLE_PGM'
-export const SET_PGM = 'SET_PGM'
-export const TOGGLE_VO = 'TOGGLE_VO'
-export const SET_VO = 'SET_VO'
-export const TOGGLE_PST = 'TOGGLE_PST'
-export const SET_PST = 'SET_PST'
-export const SET_PST_VO = 'SET_PST_VO'
-export const TOGGLE_PFL = 'TOGGLE_PFL'
-export const SET_PFL = 'SET_PFL'
-export const TOGGLE_MUTE = 'TOGGLE_MUTE'
-export const SET_MUTE = 'SET_MUTE'
-export const SHOW_CHANNEL = 'SHOW_CHANNEL'
-export const TOGGLE_SNAP = 'TOGGLE_SNAP'
-export const X_MIX = 'X_MIX'
-export const NEXT_MIX = 'NEXT_MIX'
-export const FADE_TO_BLACK = 'FADE_TO_BLACK'
-export const CLEAR_PST = 'CLEAR_PST'
-export const SNAP_RECALL = 'SNAP_RECALL'
