--- src/reducers/settingsActions.ts
@@ -1,8 +1,0 @@
-export const TOGGLE_SHOW_SETTINGS = 'TOGGLE_SHOW_SETTINGS'
-export const TOGGLE_SHOW_CHAN_STRIP = 'TOGGLE_SHOW_CHAN_STRIP'
-export const TOGGLE_SHOW_OPTION = 'TOGGLE_SHOW_OPTION'
-export const TOGGLE_SHOW_MONITOR_OPTIONS = 'TOGGLE_SHOW_MONITOR_OPTIONS'
-export const TOGGLE_SHOW_STORAGE = 'TOGGLE_SHOW_STORAGE'
-export const TOGGLE_SHOW_SNAPS = 'TOGGLE_SHOW_SNAPS'
-export const UPDATE_SETTINGS = 'UPDATE_SETTINGS'
-export const SET_MIXER_ONLINE = 'SET_MIXER_ONLINE'
