+++ webpack.build.config.js
@@ -3,7 +3,10 @@ const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 
 // Any directories you will be adding code/files into, need to be added to this array so webpack will pick them up
-const defaultInclude = path.resolve(__dirname, 'src')
+const defaultInclude = [
+  path.resolve(__dirname, 'src'),
+  path.resolve(__dirname, 'server')
+]
 
 module.exports = {
   module: {
