+++ webpack.dev.config.js
@@ -4,7 +4,11 @@ const HtmlWebpackPlugin = require('html-webpack-plugin')
 const { spawn } = require('child_process')
 
 // Any directories you will be adding code/files into, need to be added to this array so webpack will pick them up
-const defaultInclude = path.resolve(__dirname, 'src')
+const defaultInclude = [
+    path.resolve(__dirname, 'src'),
+    path.resolve(__dirname, 'server')
+]
+
 
 module.exports = {
     module: {
